// types.ts
export type Page =
  | 'home'
  | 'welcome'
  | 'image-analysis'
  | 'prescription-analysis'
  | 'checkup'
  | 'mental-health'
  | 'symptom-checker'
  | 'about'
  | 'contact'
  | 'explore'
  | 'live-alerts'
  | 'health-briefing'
  | 'activity-history'
  | 'profile'
  | 'admin-dashboard'
  | 'water-log';
// ...
export interface ChatMessage {
  id: string;
  role: 'user' | 'bot';
  text: string;
}